Simplify broadcast extension setup (#555)
The broadcast extension [in the example
app](https://github.com/livekit-examples/swift-example/tree/main/iOS/BroadcastExt)
currently defines a log handler for OSLog and bootstraps the logging
system using it. This setup allows log output to be viewed in Console
for troubleshooting the broadcast extension.

Since this configuration is generally useful across applications, I have
integrated this functionality into `LKSampleHandler` and added a
property that subclasses can override to opt in. By default, this
feature is disabled to ensure compatibility with apps that manually
bootstrap the logging system, as the example app currently does. With
this change, the boilerplate required to set up logging for the
broadcast extension is significantly reduced:
```swift
import LiveKit

#if os(iOS)
@available(macCatalyst 13.1, *)
class SampleHandler: LKSampleHandler {
    override var enableLogging: Bool { true }
}
#endif
```
If this PR is approved, I will open a separate PR updating the code in
the example app to reflect the example above.