Revamp broadcast IPC (#565)
This PR revamps inter-process communication between the main app and the
broadcast extension, replacing deprecated Core Foundation APIs.
Additionally, it enhances testability, enables bi-directional
communication, and introduces a more flexible message structure.

Summary of new types:
- `IPCChannel`: An abstraction for inter-process communication, built on
top of the [Network](https://developer.apple.com/documentation/network)
framework. Allows asynchronous sending and receiving of messages
consisting of a dynamic header (any type conforming to `Codable`) and a
data payload.
- `BroadcastImageCodec`: Encapsulates functionality for
encoding/decoding image samples for transport. For now, this uses the
same method of JPEG encoding/decoding as the current implementation.
- `BroadcastUploader`: Sends samples from ReplayKit to the main app,
built on top of `IPCChannel`.
- `BroadcastReceiver`: Receives samples from the broadcast extension,
built on top of `IPCChannel`.

This PR sets the groundwork for adding support for audio samples, but I
decided to submit that in a separate PR (#576) as this is a large change
(though not breaking).

---------

Co-authored-by: Hiroshi Horie <548776+hiroshihorie@users.noreply.github.com>