clamped the effectiveTimeout and fixed the comment (#821)
So the problem is that the performRpc is a public function, and we never
sanity check if users provide values larger than the connection_timeout,
which might cause unexpected behavior.

This PR fixes the comment and also clamp the connection response timeout
if needed.

See similar PR on JS SDK:
https://github.com/livekit/client-sdk-js/pull/1694